Got a "rough draft" of is_obstructed? implemented, added a few more tests and updated piece factory
I Got an  is_obstructed? method put togther that does successfully pass tests but seems like it could probably be cleaned up a bit. I did separate the different cases into different methods as suggested but I'm not sure if the way I went through to check if the spaces in between were occupied was the best way or not. I had two ideas, either to loop through all the pieces and see if any of them were in the way or just check the spaces along the path and see if any of them are occupied. I eventually decided to do the latter of the two since it seemed like it made more sense only to check the spaces we care about instead of looking at every piece. This method worked pretty well, I think, for horizontal and vertical paths but for diaganol paths it was a little more complicated. So not sure if maybe the other way would've been better or if I just need to rethink of how to do my loops. Along with that, there also seem like there might be a DRY issue with the way it is now since there is a lot of similar code being used in multiple spots so might need see if there is away to fix that up as well. Other than that it seems to be working at the moment so we at least have something usable that can be cleaned up if needed.

I added some more tests to to test all the different directions (left, right, up, down, up and right, up and left, down and right, down and left) and I also added a test for when the piece does not move.

Found out that I needed to capitalize the q in queen for the type of the piece so the STI would work