Added check_path method to reduce redundant code
Tried to fix some of the DRY problems mentioned before by adding a check_path method that take a row and column direction as well as distance. It then check along the path in that direction for that distance looking to see if there are any pieces along the way. Still not sure this is the best solution but I think it is getting there. :smile: