Removing some unecessary code as well as fixing some tests
I had some code I was just playing around with that wasn't really necessary for this branch so I removed it. I also removed the unecessary specs that were for the pieces show page. Also since we changed the pieces from links to spans I needed to update the game show specs.