Refactoring and bug fixes for pieces
So a littel bit about how all these changes came about, I in process of switching over to another one of my cards for this week and decided just to play around with the game. While playing around with moving pieces and such I noticed there was a bug with some of the pawn valid_move? code, more specifically I forgotten to check for any obstructions when the pawn moves forward two spaces. So I started to look into the problem and in the process decided to look as some of the other valid_moves? for other pieces to just kind of curious to see what everyone else was doing of checking for obstructions. In the process of doing this I began to see there were a few similar if not identical methods across the different pieces which seemed like a good opportunity to DRY things up little. Anyway, this kind of set me off on a testing cleaning rampage so I decided to play  around with the game a bit more to see what other things I might have overlooked, one thing lead to another and here we are with these changes

I don't want to be 'that guy', so if feel free to ignore these changes, make suggestions or cherry pick which ones you would like to use all feed back is welcome :)

**Changes:**
* DRYes up different piece type code a little bit by creating unobstructed? and unoccupied_space? and making the moving_xxlly methods protected instead of private in piece.rb so that all the other types can use them
* Refactores some of the different piece type's valid_move? methods to use the new unobstructed? and unoccupied_space? methods as well as using piece.rb versions of moving_xxlly
* Fixes problem of pawns being able to move through other pieces when trying to move forward two spaces, this is now considered in valid (as it should've been in the first place :) )
* Adds can_move_two? method to pawn.rb to clean things up a little
* Finally did some clean up on the obstructed? code since I had originally written it with the idea that (0,0) would be in the bottom left when in reality (0,0) is in the top right so had to fix a little bit of the naming and logic (probably a dumb idea to potentially break a lot of logic in diffrent places just to make it the naming is correct but if tests are to be believed, everything is still working)
* Found a odd behavior for rook and bishop, couldn't move to unoccupied spaces if the path was obstructed however, they could still capture pieces even if the path was obstructed, ended having to slightly alter the logic in those two pieces valid_move? methods
* I also added tests to rook, bishop and queen to make sure that none of them should be able to capture a piece if their path is blocked
* I also noticed that the queen couldn't move to an unoccupied location so made a slight change to that logic as well to fix that issue
* Ended making some changes to queen_spec.rb due to other changes, took out tests that were testing the helper methods that are now or already were a part of piece.rb and also added a couple new ones

Ok I think that about covers it. Again I understand I went a little crazy on this and maybe got a little ahead of myself and or oversteped my bounds so if you guys have any opinions on these changes one way or another please let me know!